[Wallet] Fix syncing issue caused by circular dependency (#1252)
