+++ packages/mobile/src/geth/networkConfig.ts
@@ -1,7 +1,21 @@
-import { DEFAULT_SYNC_MODE } from 'src/config'
-import { Testnets } from 'src/web3/testnets'
+import { DEFAULT_SYNC_MODE, DEFAULT_TESTNET } from 'src/config'
+import { GethSyncMode } from 'src/geth/consts'
 
-export default {
+export enum Testnets {
+  integration = 'integration',
+  alfajoresstaging = 'alfajoresstaging',
+  alfajores = 'alfajores',
+  pilot = 'pilot',
+  pilotstaging = 'pilotstaging',
+}
+
+interface NetworkConfig {
+  nodeDir: string
+  syncMode: GethSyncMode
+  blockchainApiUrl: string
+}
+
+const networkConfigs: { [testnet: string]: NetworkConfig } = {
   [Testnets.integration]: {
     nodeDir: `.${Testnets.integration}`,
     syncMode: DEFAULT_SYNC_MODE,
@@ -28,3 +42,5 @@ export default {
     blockchainApiUrl: 'https://pilotstaging-dot-celo-testnet.appspot.com/',
   },
 }
+
+export default networkConfigs[DEFAULT_TESTNET]
