--- packages/mobile/src/web3/testnets.ts
@@ -1,7 +1,0 @@
-export enum Testnets {
-  integration = 'integration',
-  alfajoresstaging = 'alfajoresstaging',
-  alfajores = 'alfajores',
-  pilot = 'pilot',
-  pilotstaging = 'pilotstaging',
-}
